Extract helper functions to own module file
